[feat]微服务改造，api网关、account，rpc调用，暂存

+++ pkg/server/result.go
@@ -1,7 +1,7 @@
 package server
 
 type Result struct {
-	Code int    `json:"code,omitempty"`
+	Code int64  `json:"code,omitempty"`
 	Msg  string `json:"msg,omitempty"`
 	Data any    `json:"data,omitempty"`
 }
